+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Implemented `Clock` dependency.
 - Improved handling of file parameters for `FileSystem::Mock`.
 - Fixed a `FileSystem::Mock` bug that would wipe existing mock directories.
+- Fixed a `FileSystem::Mock` bug that did not include directories in `each_file`.
 
 ## Files